+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2024-01-08
+
+### Added
+ - Integration of Physical Memory Attribute (PMA) modules
+
+### Changed
+ - Propagation of correctable errors at fetch interface from ID to MA stage
+ - IMISCON and FETCH error definitions rearranged
+ - IMISCON_PRED was removed and merged with IMISCON_DSCR
+
+### Removed
+ - Possibility to configure protected core without interface protection
+ - Checking of the FETCH address discrepancy since it is not needed anymore (after integration of bus protection)
+ - Boot address as a post-compilation option
+
+### Fixed
+ - ACM reliability issue
+
 ## [1.2.2] - 2024-01-05
 
 ### Added
@@ -96,6 +114,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial public version
 
+[1.3.0]: https://github.com/janomach/the-hardisc/releases/tag/v1.3.0
 [1.2.2]: https://github.com/janomach/the-hardisc/releases/tag/v1.2.2
 [1.2.1]: https://github.com/janomach/the-hardisc/releases/tag/v1.2.1
 [1.2.0]: https://github.com/janomach/the-hardisc/releases/tag/v1.2.0